requirements.txt
ruamel.yaml
theblues
libcharmstore
git+https://github.com/openstack-charmers/charmed-openstack-info.git#egg=charmed-openstack-info
GitPython
humanize
launchpadlib
